# feat(col.bool): improve performance