```
name                     old time/op    new time/op    delta
ColBool_DecodeColumn-32    23.7µs ± 1%    23.2µs ± 2%     ~     (p=0.095 n=5+5)
ColBool_EncodeColumn-32    28.0µs ± 2%    21.9µs ± 2%  -21.72%  (p=0.008 n=5+5)

name                     old speed      new speed      delta
ColBool_DecodeColumn-32  2.11GB/s ± 1%  2.16GB/s ± 2%     ~     (p=0.095 n=5+5)
ColBool_EncodeColumn-32  1.79GB/s ± 2%  2.28GB/s ± 2%  +27.73%  (p=0.008 n=5+5)
```

kudos to @cristaloleg